New file structure ver 0.8.7.5
Nearly finish mesh asset save load

+++ Engine/Sources/EngineCore/Core/PigeonBase/Object/World/World.cpp
@@ -29,7 +29,7 @@ namespace PigeonEngine
         Check(ENGINE_WORLD_ERROR, "PWorld::AddActor : Adding nullptr to world.", GameTimer == nullptr);
         for(const auto& Actor : AllActors)
         {
-            Actor->FixedTick(GameTimer->GetDeltaTime());
+            Actor->FixedTick(static_cast<FLOAT>(GameTimer->GetDeltaTime()));
         }
     }
